\name{AQuality-package}
\alias{AQuality-package}
\alias{AQuality}
\docType{package}
\title{
\packageTitle{AQuality}
}
\description{The package allows you tu evaluate graphically the quality of measurements of water components}
\details{The package includes two functions: TSSS() and ICHS(). The TSSS()function  allows evaluating the quality of a set of measurement of water components, which correlate with total soluble solids. On the other hand, the ICHS() function allows evaluating the quality of a set of measurement of water components, which correlate with conductivity}
\author{
\packageAuthor{AQuality}
Maintainer: Maela Lupo email: maela.lupo@gmail.com}
\keyword{package}