\name{AQuality-package}
\alias{AQuality-package}
\alias{AQuality}
\docType{package}
\title{
\packageTitle{AQuality}
}
\description{The package allows you tu evaluate graphically the quality of measurements of water components}
\details{The package includes the TSSS() function that allows evaluating the quality of a set of measurement of water components, which correlate with total soluble solids}
\author{
\packageAuthor{AQuality}
Maintainer: Maela Lupo email: maela.lupo@gmail.com}
\keyword{package}